Extract metrics.py and results.py from backtest.py
Reduces backtest.py from 1375 to 1025 lines by pulling out two
independent concerns:

- metrics.py: pure performance-metric functions (Sharpe, Sortino, CAGR,
  drawdown, trade stats) and StrategyStats dataclass — no engine
  dependency, importable by any consumer.
- results.py: BacktestResult dataclass and CSV/JSON serialization
  helpers — depends on metrics, not on the engine.

Also standardises module-level declaration order (constants before
dataclasses before functions) in models.py, config.py, and output.py.
